% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-logsumexp.R
\name{eb_log_sum_exp}
\alias{eb_log_sum_exp}
\title{Compute a Stable Log-Sum-Exp}
\usage{
eb_log_sum_exp(x)
}
\arguments{
\item{x}{A numeric vector on the log scale.}
}
\value{
A numeric scalar.
}
\description{
Numerically stable scalar log-sum-exp helper.
}
