% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot_eb_estimates}
\alias{plot_eb_estimates}
\alias{plot.eb_estimates}
\title{Base plotting for \code{eb_estimates} objects}
\usage{
\method{plot}{eb_estimates}(x, y = NULL, type = c("histogram", "qq"), ...)
}
\arguments{
\item{x}{An \code{eb_estimates} object.}

\item{y}{Unused.}

\item{type}{Plot type to construct.}

\item{...}{Additional graphical arguments passed to the underlying base plot.}
}
\value{
The input object, invisibly.
}
\description{
\code{plot.eb_estimates()} helps inspect the observed estimate layer before
shrinkage. Use \code{"histogram"} for the marginal estimate distribution and
\code{"qq"} for a normal QQ check.
}
\examples{
est <- eb_input(
  theta_hat = c(-0.10, 0.05, 0.20, 0.35),
  s = c(0.20, 0.20, 0.20, 0.20)
)

plot(est)
plot(est, type = "qq")

}
