% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{plot_eb_sim}
\alias{plot_eb_sim}
\alias{plot.eb_sim}
\title{Base plotting for \code{eb_sim} objects}
\usage{
\method{plot}{eb_sim}(x, y = NULL, type = c("truth", "density"), ...)
}
\arguments{
\item{x}{An \code{eb_sim} object.}

\item{y}{Unused.}

\item{type}{Plot type to construct.}

\item{...}{Additional graphical arguments passed to the underlying base plot.}
}
\value{
The input object, invisibly.
}
\description{
\code{plot.eb_sim()} provides quick views of the simulated truth:
}
\details{
\itemize{
\item \code{"truth"} plots unit-level true effects in index order
\item \code{"density"} plots the empirical density of the true effects
}
}
\examples{
sim <- eb_simulate(n_units = 8, n_obs = 80, seed = 1)

plot(sim)
plot(sim, type = "density")

}
